feat: added balance check to precheck
