+++ pallets/sponsored-transactions/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pallet-meta-transactions"
+name = "pallet-sponsored-transactions"
 version = "0.1.0"
 edition = "2021"
