+++ pallets/sponsored-transactions/src/mock.rs
@@ -1,0 +1,274 @@
+#![cfg(test)]
+
+use core::str::FromStr;
+
+use runner::Runner as StabilityRunner;
+
+use frame_support::{
+	construct_runtime,
+	pallet_prelude::{StorageValue, ValueQuery},
+	parameter_types,
+	traits::{Everything, GenesisBuild, StorageInstance},
+	weights::Weight,
+};
+use pallet_evm::{EnsureAddressNever, EnsureAddressRoot};
+use sp_core::{H160, H256, U256};
+use sp_runtime::{
+	traits::{BlakeTwo256, IdentifyAccount, IdentityLookup, Verify},
+	MultiSignature,
+};
+use std::collections::BTreeMap;
+
+pub type Signature = MultiSignature;
+
+pub type AccountId = <<Signature as Verify>::Signer as IdentifyAccount>::AccountId;
+pub type Balance = u128;
+pub type BlockNumber = u32;
+type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Runtime>;
+type Block = frame_system::mocking::MockBlock<Runtime>;
+
+pub struct MockUserFeeTokenController;
+impl pallet_user_fee_selector::UserFeeTokenController for MockUserFeeTokenController {
+	type Error = ();
+
+	fn balance_of(_token: H160) -> sp_core::U256 {
+		return U256::from(1000000);
+	}
+
+	fn get_user_fee_token(_account: H160) -> H160 {
+		return ERC20SlotZero::get();
+	}
+
+	fn set_user_fee_token(_account: H160, _token: H160) -> Result<(), Self::Error> {
+		Ok(())
+	}
+}
+
+parameter_types! {
+	pub const BlockHashCount: u32 = 250;
+	pub const SS58Prefix: u8 = 42;
+}
+
+impl frame_system::Config for Runtime {
+	type BaseCallFilter = Everything;
+	type DbWeight = ();
+	type RuntimeOrigin = RuntimeOrigin;
+	type Index = u64;
+	type BlockNumber = BlockNumber;
+	type RuntimeCall = RuntimeCall;
+	type Hash = H256;
+	type Hashing = BlakeTwo256;
+	type AccountId = AccountId;
+	type Lookup = IdentityLookup<AccountId>;
+	type Header = sp_runtime::generic::Header<BlockNumber, BlakeTwo256>;
+	type RuntimeEvent = RuntimeEvent;
+	type BlockHashCount = BlockHashCount;
+	type Version = ();
+	type PalletInfo = PalletInfo;
+	type AccountData = pallet_balances::AccountData<Balance>;
+	type OnNewAccount = ();
+	type OnKilledAccount = ();
+	type SystemWeightInfo = ();
+	type BlockWeights = ();
+	type BlockLength = ();
+	type SS58Prefix = SS58Prefix;
+	type OnSetCode = ();
+	type MaxConsumers = frame_support::traits::ConstU32<16>;
+}
+
+parameter_types! {
+	pub const ExistentialDeposit: u128 = 0;
+}
+
+impl pallet_balances::Config for Runtime {
+	type MaxReserves = ();
+	type ReserveIdentifier = ();
+	type MaxLocks = ();
+	type Balance = Balance;
+	type RuntimeEvent = RuntimeEvent;
+	type DustRemoval = ();
+	type ExistentialDeposit = ExistentialDeposit;
+	type AccountStore = System;
+	type WeightInfo = ();
+}
+
+parameter_types! {
+	pub const MinimumPeriod: u64 = 5;
+}
+
+impl pallet_timestamp::Config for Runtime {
+	type Moment = u64;
+	type OnTimestampSet = ();
+	type MinimumPeriod = MinimumPeriod;
+	type WeightInfo = ();
+}
+
+pub struct MockDNTFeeController;
+impl runner::OnChargeDecentralizedNativeTokenFee for MockDNTFeeController {
+	type Error = ();
+
+	fn get_transaction_fee_token(_from: H160) -> H160 {
+		Default::default()
+	}
+
+	fn get_transaction_conversion_rate(_validator: H160, _token: H160) -> (U256, U256) {
+		(1.into(), 1.into())
+	}
+
+	fn get_fee_vault() -> H160 {
+		Default::default()
+	}
+
+	fn withdraw_fee(
+		_from: H160,
+		_token: H160,
+		_conversion_rate: (U256, U256),
+		_amount: U256,
+	) -> Result<(), Self::Error> {
+		Ok(())
+	}
+
+	fn correct_fee(
+		_from: H160,
+		_token: H160,
+		_conversion_rate: (U256, U256),
+		_paid_amount: U256,
+		_actual_amount: U256,
+	) -> Result<(), Self::Error> {
+		Ok(())
+	}
+
+	fn pay_fees(
+		_token: H160,
+		_conversion_rate: (U256, U256),
+		_actual_amount: U256,
+		_validator: H160,
+		_to: Option<H160>,
+	) -> Result<(U256, U256), Self::Error> {
+		Ok((Default::default(), Default::default()))
+	}
+}
+
+parameter_types! {
+	pub BlockGasLimit: U256 = U256::max_value();
+	pub const WeightPerGas: Weight = Weight::from_ref_time(1);
+	pub ERC20SlotZero: H160 = H160::from_str("0x22D598E0a9a1b474CdC7c6fBeA0B4F83E12046a9").unwrap();
+	pub ZeroSlot : H256 = H256::from_low_u64_be(0);
+}
+
+parameter_types! {
+	pub const ChainId: u64 = 20180428;
+}
+impl pallet_evm::Config for Runtime {
+	type FeeCalculator = ();
+	type GasWeightMapping = pallet_evm::FixedGasWeightMapping<Self>;
+	type WeightPerGas = WeightPerGas;
+	type CallOrigin = EnsureAddressRoot<AccountId>;
+	type WithdrawOrigin = EnsureAddressNever<AccountId>;
+	type AddressMapping = pallet_evm::HashedAddressMapping<BlakeTwo256>;
+	type Currency = Balances;
+	type RuntimeEvent = RuntimeEvent;
+	type Runner = StabilityRunner<Self, MockDNTFeeController, MockUserFeeTokenController>;
+	type PrecompilesType = ();
+	type PrecompilesValue = ();
+	type ChainId = ChainId;
+	type OnChargeTransaction = ();
+	type BlockGasLimit = BlockGasLimit;
+	type BlockHashMapping = pallet_evm::SubstrateBlockHashMapping<Self>;
+	type FindAuthor = ();
+}
+
+impl pallet_ethereum::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+
+	type StateRoot = ();
+}
+
+pub struct MockPrefix;
+impl StorageInstance for MockPrefix {
+	fn pallet_prefix() -> &'static str {
+		"MockPrefix"
+	}
+
+	const STORAGE_PREFIX: &'static str = "MockPrefix";
+}
+
+// Bool value is true for deposit and false for withdraw
+pub type StorageCalledArguments =
+	StorageValue<MockPrefix, Vec<(bool, H160, H160, U256)>, ValueQuery>;
+
+pub struct MockERC20Manager;
+impl pallet_erc20_manager::ERC20Manager for MockERC20Manager {
+	type Error = ();
+
+	fn balance_of(_token: H160, _payer: H160) -> U256 {
+		U256::max_value()
+	}
+
+	fn withdraw_amount(token: H160, from: H160, amount: U256) -> Result<U256, Self::Error> {
+		StorageCalledArguments::mutate(|args| {
+			args.push((false, token, from, amount));
+		});
+		Ok(amount.into())
+	}
+
+	fn deposit_amount(token: H160, from: H160, amount: U256) -> Result<U256, Self::Error> {
+		StorageCalledArguments::mutate(|args| {
+			args.push((true, token, from, amount));
+		});
+		Ok(amount.into())
+	}
+}
+
+impl crate::Config for Runtime {
+	type RuntimeCall = RuntimeCall;
+	type ERC20Manager = MockERC20Manager;
+	type DNTFeeController = MockDNTFeeController;
+}
+
+// Configure a mock runtime to test the pallet.
+construct_runtime!(
+	pub enum Runtime where
+		Block = Block,
+		NodeBlock = Block,
+		UncheckedExtrinsic = UncheckedExtrinsic,
+	{
+		System: frame_system::{Pallet, Call, Config, Storage, Event<T>},
+		Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
+		Evm: pallet_evm::{Pallet, Call, Storage, Event<T>},
+		Timestamp: pallet_timestamp::{Pallet, Call, Storage, Inherent},
+		Ethereum: pallet_ethereum,
+		MetaTransactions: crate
+	}
+);
+
+parameter_types! {
+	pub MetaDeploymentTransaction: Vec<u8> = hex::decode("02f921cc840133edcc80849502f90085010c388d00832dc6c08080b9216e60806040523480156200001157600080fd5b506040518060400160405280600f81526020017f4d65616e696e676c6573732045555200000000000000000000000000000000008152506040518060400160405280600481526020017f4d4555520000000000000000000000000000000000000000000000000000000081525081600390805190602001906200009692919062000357565b508060049080519060200190620000af92919062000357565b505050620000d2620000c6620000fb60201b60201c565b6200010360201b60201c565b620000f5337002f050fe938943acc45f65568000000000620001c960201b60201c565b620005b3565b600033905090565b6000600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905081600560006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b620001db8282620001df60201b60201c565b5050565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff16141562000252576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040162000249906200043f565b60405180910390fd5b62000266600083836200034d60201b60201c565b80600260008282546200027a91906200048f565b92505081905550806000808473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508173ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef836040516200032d919062000461565b60405180910390a362000349600083836200035260201b60201c565b5050565b505050565b505050565b8280546200036590620004f6565b90600052602060002090601f016020900481019282620003895760008555620003d5565b82601f10620003a457805160ff1916838001178555620003d5565b82800160010185558215620003d5579182015b82811115620003d4578251825591602001919060010190620003b7565b5b509050620003e49190620003e8565b5090565b5b8082111562000403576000816000905550600101620003e9565b5090565b600062000416601f836200047e565b915062000423826200058a565b602082019050919050565b6200043981620004ec565b82525050565b600060208201905081810360008301526200045a8162000407565b9050919050565b60006020820190506200047860008301846200042e565b92915050565b600082825260208201905092915050565b60006200049c82620004ec565b9150620004a983620004ec565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff03821115620004e157620004e06200052c565b5b828201905092915050565b6000819050919050565b600060028204905060018216806200050f57607f821691505b602082108114156200052657620005256200055b565b5b50919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b7f45524332303a206d696e7420746f20746865207a65726f206164647265737300600082015250565b611bab80620005c36000396000f3fe608060405234801561001057600080fd5b506004361061010b5760003560e01c806370a08231116100a257806395d89b411161007157806395d89b41146102a6578063a457c2d7146102c4578063a9059cbb146102f4578063dd62ed3e14610324578063f2fde38b146103545761010b565b806370a0823114610232578063715018a61461026257806379cc67901461026c5780638da5cb5b146102885761010b565b8063313ce567116100de578063313ce567146101ac57806339509351146101ca57806340c10f19146101fa57806342966c68146102165761010b565b806306fdde0314610110578063095ea7b31461012e57806318160ddd1461015e57806323b872dd1461017c575b600080fd5b610118610370565b604051610125919061148d565b60405180910390f35b610148600480360381019061014391906111d1565b610402565b6040516101559190611472565b60405180910390f35b610166610425565b604051610173919061162f565b60405180910390f35b6101966004803603810190610191919061117e565b61042f565b6040516101a39190611472565b60405180910390f35b6101b461045e565b6040516101c1919061164a565b60405180910390f35b6101e460048036038101906101df91906111d1565b610467565b6040516101f19190611472565b60405180910390f35b610214600480360381019061020f91906111d1565b61049e565b005b610230600480360381019061022b9190611211565b6104ac565b005b61024c60048036038101906102479190611111565b6104c0565b604051610259919061162f565b60405180910390f35b61026a610508565b005b610286600480360381019061028191906111d1565b61051c565b005b61029061053c565b60405161029d9190611457565b60405180910390f35b6102ae610566565b6040516102bb919061148d565b60405180910390f35b6102de60048036038101906102d991906111d1565b6105f8565b6040516102eb9190611472565b60405180910390f35b61030e600480360381019061030991906111d1565b61066f565b60405161031b9190611472565b60405180910390f35b61033e6004803603810190610339919061113e565b610692565b60405161034b919061162f565b60405180910390f35b61036e60048036038101906103699190611111565b610719565b005b60606003805461037f9061175f565b80601f01602080910402602001604051908101604052809291908181526020018280546103ab9061175f565b80156103f85780601f106103cd576101008083540402835291602001916103f8565b820191906000526020600020905b8154815290600101906020018083116103db57829003601f168201915b5050505050905090565b60008061040d61079d565b905061041a8185856107a5565b600191505092915050565b6000600254905090565b60008061043a61079d565b9050610447858285610970565b6104528585856109fc565b60019150509392505050565b60006012905090565b60008061047261079d565b90506104938185856104848589610692565b61048e9190611681565b6107a5565b600191505092915050565b6104a88282610c74565b5050565b6104bd6104b761079d565b82610dcb565b50565b60008060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020549050919050565b610510610f99565b61051a6000611017565b565b61052e8261052861079d565b83610970565b6105388282610dcb565b5050565b6000600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905090565b6060600480546105759061175f565b80601f01602080910402602001604051908101604052809291908181526020018280546105a19061175f565b80156105ee5780601f106105c3576101008083540402835291602001916105ee565b820191906000526020600020905b8154815290600101906020018083116105d157829003601f168201915b5050505050905090565b60008061060361079d565b905060006106118286610692565b905083811015610656576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161064d906115ef565b60405180910390fd5b61066382868684036107a5565b60019250505092915050565b60008061067a61079d565b90506106878185856109fc565b600191505092915050565b6000600160008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008373ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905092915050565b610721610f99565b600073ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff161415610791576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610788906114ef565b60405180910390fd5b61079a81611017565b50565b600033905090565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff161415610815576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161080c906115cf565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff161415610885576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161087c9061150f565b60405180910390fd5b80600160008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020819055508173ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167f8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b92583604051610963919061162f565b60405180910390a3505050565b600061097c8484610692565b90507fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff81146109f657818110156109e8576040517f08c379a00000000000000000000000000000000000000000000000000000000081526004016109df9061152f565b60405180910390fd5b6109f584848484036107a5565b5b50505050565b600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff161415610a6c576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610a63906115af565b60405180910390fd5b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff161415610adc576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610ad3906114af565b60405180910390fd5b610ae78383836110dd565b60008060008573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905081811015610b6d576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610b649061154f565b60405180910390fd5b8181036000808673ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002081905550816000808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508273ffffffffffffffffffffffffffffffffffffffff168473ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef84604051610c5b919061162f565b60405180910390a3610c6e8484846110e2565b50505050565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff161415610ce4576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610cdb9061160f565b60405180910390fd5b610cf0600083836110dd565b8060026000828254610d029190611681565b92505081905550806000808473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff168152602001908152602001600020600082825401925050819055508173ffffffffffffffffffffffffffffffffffffffff16600073ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef83604051610db3919061162f565b60405180910390a3610dc7600083836110e2565b5050565b600073ffffffffffffffffffffffffffffffffffffffff168273ffffffffffffffffffffffffffffffffffffffff161415610e3b576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610e329061158f565b60405180910390fd5b610e47826000836110dd565b60008060008473ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff16815260200190815260200160002054905081811015610ecd576040517f08c379a0000000000000000000000000000000000000000000000000000000008152600401610ec4906114cf565b60405180910390fd5b8181036000808573ffffffffffffffffffffffffffffffffffffffff1673ffffffffffffffffffffffffffffffffffffffff1681526020019081526020016000208190555081600260008282540392505081905550600073ffffffffffffffffffffffffffffffffffffffff168373ffffffffffffffffffffffffffffffffffffffff167fddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef84604051610f80919061162f565b60405180910390a3610f94836000846110e2565b505050565b610fa161079d565b73ffffffffffffffffffffffffffffffffffffffff16610fbf61053c565b73ffffffffffffffffffffffffffffffffffffffff1614611015576040517f08c379a000000000000000000000000000000000000000000000000000000000815260040161100c9061156f565b60405180910390fd5b565b6000600560009054906101000a900473ffffffffffffffffffffffffffffffffffffffff16905081600560006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055508173ffffffffffffffffffffffffffffffffffffffff168173ffffffffffffffffffffffffffffffffffffffff167f8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e060405160405180910390a35050565b505050565b505050565b6000813590506110f681611b47565b92915050565b60008135905061110b81611b5e565b92915050565b600060208284031215611127576111266117ef565b5b6000611135848285016110e7565b91505092915050565b60008060408385031215611155576111546117ef565b5b6000611163858286016110e7565b9250506020611174858286016110e7565b9150509250929050565b600080600060608486031215611197576111966117ef565b5b60006111a5868287016110e7565b93505060206111b6868287016110e7565b92505060406111c7868287016110fc565b9150509250925092565b600080604083850312156111e8576111e76117ef565b5b60006111f6858286016110e7565b9250506020611207858286016110fc565b9150509250929050565b600060208284031215611227576112266117ef565b5b6000611235848285016110fc565b91505092915050565b611247816116d7565b82525050565b611256816116e9565b82525050565b600061126782611665565b6112718185611670565b935061128181856020860161172c565b61128a816117f4565b840191505092915050565b60006112a2602383611670565b91506112ad82611805565b604082019050919050565b60006112c5602283611670565b91506112d082611854565b604082019050919050565b60006112e8602683611670565b91506112f3826118a3565b604082019050919050565b600061130b602283611670565b9150611316826118f2565b604082019050919050565b600061132e601d83611670565b915061133982611941565b602082019050919050565b6000611351602683611670565b915061135c8261196a565b604082019050919050565b6000611374602083611670565b915061137f826119b9565b602082019050919050565b6000611397602183611670565b91506113a2826119e2565b604082019050919050565b60006113ba602583611670565b91506113c582611a31565b604082019050919050565b60006113dd602483611670565b91506113e882611a80565b604082019050919050565b6000611400602583611670565b915061140b82611acf565b604082019050919050565b6000611423601f83611670565b915061142e82611b1e565b602082019050919050565b61144281611715565b82525050565b6114518161171f565b82525050565b600060208201905061146c600083018461123e565b92915050565b6000602082019050611487600083018461124d565b92915050565b600060208201905081810360008301526114a7818461125c565b905092915050565b600060208201905081810360008301526114c881611295565b9050919050565b600060208201905081810360008301526114e8816112b8565b9050919050565b60006020820190508181036000830152611508816112db565b9050919050565b60006020820190508181036000830152611528816112fe565b9050919050565b6000602082019050818103600083015261154881611321565b9050919050565b6000602082019050818103600083015261156881611344565b9050919050565b6000602082019050818103600083015261158881611367565b9050919050565b600060208201905081810360008301526115a88161138a565b9050919050565b600060208201905081810360008301526115c8816113ad565b9050919050565b600060208201905081810360008301526115e8816113d0565b9050919050565b60006020820190508181036000830152611608816113f3565b9050919050565b6000602082019050818103600083015261162881611416565b9050919050565b60006020820190506116446000830184611439565b92915050565b600060208201905061165f6000830184611448565b92915050565b600081519050919050565b600082825260208201905092915050565b600061168c82611715565b915061169783611715565b9250827fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff038211156116cc576116cb611791565b5b828201905092915050565b60006116e2826116f5565b9050919050565b60008115159050919050565b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000819050919050565b600060ff82169050919050565b60005b8381101561174a57808201518184015260208101905061172f565b83811115611759576000848401525b50505050565b6000600282049050600182168061177757607f821691505b6020821081141561178b5761178a6117c0565b5b50919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052601160045260246000fd5b7f4e487b7100000000000000000000000000000000000000000000000000000000600052602260045260246000fd5b600080fd5b6000601f19601f8301169050919050565b7f45524332303a207472616e7366657220746f20746865207a65726f206164647260008201527f6573730000000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a206275726e20616d6f756e7420657863656564732062616c616e60008201527f6365000000000000000000000000000000000000000000000000000000000000602082015250565b7f4f776e61626c653a206e6577206f776e657220697320746865207a65726f206160008201527f6464726573730000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a20617070726f766520746f20746865207a65726f20616464726560008201527f7373000000000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a20696e73756666696369656e7420616c6c6f77616e6365000000600082015250565b7f45524332303a207472616e7366657220616d6f756e742065786365656473206260008201527f616c616e63650000000000000000000000000000000000000000000000000000602082015250565b7f4f776e61626c653a2063616c6c6572206973206e6f7420746865206f776e6572600082015250565b7f45524332303a206275726e2066726f6d20746865207a65726f2061646472657360008201527f7300000000000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a207472616e736665722066726f6d20746865207a65726f20616460008201527f6472657373000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a20617070726f76652066726f6d20746865207a65726f2061646460008201527f7265737300000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a2064656372656173656420616c6c6f77616e63652062656c6f7760008201527f207a65726f000000000000000000000000000000000000000000000000000000602082015250565b7f45524332303a206d696e7420746f20746865207a65726f206164647265737300600082015250565b611b50816116d7565b8114611b5b57600080fd5b50565b611b6781611715565b8114611b7257600080fd5b5056fea2646970667358221220d51b9af5425d2e17df163956cba2377139449296789bd33fef598f221778094364736f6c63430008070033c080a0dd6784e47dc289a9afa9443ba2b4aa3a33152f78a8bf7110352541e1884994eca04e05fa49ec96a2942dba70600d0cec1e7a73cfcd43f04ab51c4bc185abdb09d9").unwrap();
+	pub MetaDeploymentSignature: Vec<u8> = hex::decode("9427c4499475af94bc06b01b10cc8df9a695469811e108fbc7ed9fe192f79c0b4d3bb33a693b2f6695acd6cb8a543b27ac7f2e10f8d3ac5e07904eb60d536ac21b").unwrap();
+	pub RawTransaction0: Vec<u8> = hex::decode("f86b80843b9aca0082520894f0a57f274781b0ed17f2ae5f1709a0c360cbe489880de0b6b3a7640000801ba0e4cb4bbd414f044046b741e5ec6d874ca73a178e6e84e5fd2b6e2f6d0bcac869a05f334d87f56a77435a2547188a935a37358b5e36e76789ac5278bd0963a931ae").unwrap();
+	pub RawTransaction1: Vec<u8> = hex::decode("f86b01843b9aca0082520894f0a57f274781b0ed17f2ae5f1709a0c360cbe489880de0b6b3a7640000801ca029343cf22be72d879c6392a2ff5dd7e3675b6b70ffc7b7a15529ed9bdad33668a04d5de0209d299c37af1642ea17ca43fdcf0b07d4aa44fe87e338b4fe9b2fa995").unwrap();
+	pub MetaTransaction0Signature: Vec<u8> = hex::decode("bce0eab34bf451c7ab98ecab110220a77f8d15b1d3c641cb02eb85ae79dddec9284b01839010ffc65dff384b17676bb1875dda86cb313b9bdef1146f2b0fd17f1c").unwrap();
+	pub MetaTransaction1Signature: Vec<u8> = hex::decode("fb0482cfa593aa03aaa1d7c2e5a07269d65a4a589b14c109de8d8ca0c268efca198d1d898e7022002e34d91ced41deda8ae526856a8e23d100ab92e5dbb948c31b").unwrap();
+	pub Sponsor: H160 = H160::from_str("0xaf537bd156c7e548d0bf2cd43168dabf7af2feb5").unwrap();
+}
+
+pub fn new_test_ext() -> sp_io::TestExternalities {
+	let mut t = frame_system::GenesisConfig::default()
+		.build_storage::<Runtime>()
+		.unwrap();
+	let config = pallet_evm::GenesisConfig {
+		accounts: BTreeMap::new(),
+	};
+	<pallet_evm::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(&config, &mut t)
+		.unwrap();
+
+	let eth_config: pallet_ethereum::GenesisConfig = Default::default();
+
+	<pallet_ethereum::GenesisConfig as GenesisBuild<Runtime>>::assimilate_storage(
+		&eth_config,
+		&mut t,
+	)
+	.unwrap();
+
+	t.into()
+}
