+++ runtime/Cargo.toml
@@ -136,7 +136,7 @@ pallet-evm-precompile-modexp = { workspace = true }
 pallet-evm-precompile-sha3fips = { workspace = true }
 pallet-evm-precompile-simple = { workspace = true }
 pallet-hotfix-sufficients = { workspace = true }
-pallet-meta-transactions = { workspace = true }
+pallet-sponsored-transactions = { workspace = true }
 rpc-eth-extension-api = { workspace = true }
 # Stability
 pallet-evm-precompile-blake2 = { workspace = true }
