+++ stability-rpc/Cargo.toml
@@ -25,7 +25,7 @@ async-trait = { workspace = true }
 # local packages
 stability-rpc-api = { workspace = true, default-features = false }
 sc-transaction-pool-api = { workspace = true }
-pallet-meta-transactions = { workspace = true }
+pallet-sponsored-transactions = { workspace = true }
 rpc-eth-extension-api = { workspace = true }
 futures-util = { version = "0.3.17" }
 jsonrpsee-core = { version = "0.18.1" }
@@ -35,7 +35,7 @@ assert_matches = "1.3.0"
 tokio = "1.17.0"
 substrate-test-runtime-client = { workspace = true }
 sc-client-api = { workspace = true }
-pallet-meta-transactions = { workspace = true }
+pallet-sponsored-transactions = { workspace = true }
 
 
 [features]
@@ -45,6 +45,6 @@ std = [
   "sp-core/std",
   "sp-runtime/std",
   "stability-rpc-api/std",
-  "pallet-meta-transactions/std"
+  "pallet-sponsored-transactions/std"
 ]
 test-helpers = ["lazy_static", "tokio"]
